Added branch with Netbeans project folder of my GUI ideas By no means is this what I think it should look like but it could be a start
After opening the project early_assessment_gui netbeans project
folder you may need to go to project properties and add the EarlyAssessmentSRC
folder as a source folder to the netbeans project for it to compile.

Let me know if you have problems running it. My email is in my github profile